+++ extensions/git-branches-panel/src/treeProvider.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 
 import { type BranchInfo, type BranchSortOrder, type BranchTreeNode } from './branchModel';
-import { fetchRemoteState, getBranches, getRemoteBranches, getRepoRoot, getTags } from './git';
+import { fetchRemoteState, getBranches, getRemoteBranches, getRepoRoot, getStashes, getTags } from './git';
 import { BranchDataLoader, type BranchDataLoaderDependencies, type BranchLoadOptions } from './treeDataLoader';
 import { BranchTreeItem } from './treeItem';
 import { buildStatusBarText, buildStatusBarTooltipContent, findContainerNode } from './treePresentation';
@@ -102,6 +102,7 @@ function createBranchDataLoaderDependencies(): BranchDataLoaderDependencies {
     getRepoRoot,
     getBranches,
     getRemoteBranches,
+    getStashes,
     getTags,
     fetchRemoteState,
     warn: (message) => {