+++ extensions/git-branches-panel/test/runtimeModules.test.js
@@ -223,6 +223,7 @@ test('BranchTreeProvider refreshes status bar data and exposes tree items', asyn
       getBranches() {},
       getRemoteBranches() {},
       getRepoRoot() {},
+      getStashes() {},
       getTags() {},
     },
     './treeDataLoader': {
@@ -304,6 +305,7 @@ test('BranchTreeProvider loads nested container children and hides the status ba
       getBranches() {},
       getRemoteBranches() {},
       getRepoRoot() {},
+      getStashes() {},
       getTags() {},
     },
     './treeDataLoader': {